feat(config): set Keystone secureCookies to false
The login to "admin" UI was impossible...
(secureCookies: false, => needed on non-SSL servers)
https://www.keystonejs.com/keystonejs/keystone/

A secure cookie is only sent to the server with an encrypted request over the HTTPS protocol. If secureCookies is set to true (as is the default with a production build) for a KeystoneJS project running on a non-HTTPS server (such as localhost), you will not be able to log in. In that case, be sure you set secureCookies to false. This does not affect development builds since this value is already false.